+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Release Notes for Feed Me 4.x
 
+### Fixed
+- Fixed a bug where importing data to the commerce variant table field don’t work. ([#464](https://github.com/craftcms/feed-me/issues/464))
+
 ## 4.6.1 - 2023-03-20
 
 - Fixed a JavaScript error that would occur on case-sensitive filesystems when using Feed Me. ([#1260](https://github.com/craftcms/feed-me/pull/1260), [#1257](https://github.com/craftcms/feed-me/issues/1257), [#1258](https://github.com/craftcms/feed-me/issues/1258), [#1259](https://github.com/craftcms/feed-me/issues/1259))